Move ChangeTeamDetailsPage up a level to match route hierarchy
