+++ ui/src/App/ChangeTeamDetails/ChangeTeamDetailsPage.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@use '../../../Styles/main';
+@use '../../Styles/main';
 
 .change-team-details-page {
 	.change-team-details-form {
